Fix #7535 - [Bug] HashCode not working on ingList

+++ becpg-core/src/main/java/fr/becpg/repo/repository/model/BeCPGDataObject.java
@@ -52,14 +52,14 @@ public  abstract class  BeCPGDataObject  extends BaseObject implements Repositor
 	
 	protected transient boolean isTransient = false;
 	
-	private transient int dbHashCode;
+	private transient long dbHashCode;
 	
 	@InternalField
-	public int getDbHashCode() {
+	public long getDbHashCode() {
 		return dbHashCode;
 	}
 
-	public void setDbHashCode(int dbHashCode) {
+	public void setDbHashCode(long dbHashCode) {
 		this.dbHashCode = dbHashCode;
 	}
